pyproject.toml
[project]
name = "soapvoice"
version = "0.8.0"
description = "醫療語音轉 SOAP 病歷系統"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "faster-whisper>=1.0.0",
    "torch>=2.1.0",
    "torchaudio>=2.1.0",
    "pyyaml>=6.0.0",
    "openai>=1.10.0",
    "redis>=5.0.0",
    "httpx>=0.25.0",
    "websockets>=12.0",
    "requests>=2.31.0",
    "pyjwt>=2.8.0",
    "pyaudio>=0.2.13",
    "scipy>=1.17.1",
    "moonshine-voice>=0.0.49",
    "funasr",
    "modelscope",
    "qwen-asr>=0.0.6",
    "opencc>=0.2.0",
    "python-docx>=1.2.0",
    "chromadb>=1.5.5",
    "sentence-transformers>=5.3.0",
    "llama-cpp-python>=0.3.18",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "locust>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "locust>=2.0.0",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

# 測試覆蓋率配置
[tool.coverage.run]
branch = true
source = ["src"]
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/conftest.py",
    "*/main.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
fail_under = 80
show_missing = true
skip_covered = false

[tool.coverage.html]
directory = "htmlcov"
title = "SoapVoice 測試覆蓋率報告"

[tool.coverage.xml]
output = "coverage.xml"
